+++ packages/devtools-client/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @prototype/devtools-client
+# prototype-devtools-client
 
 ## 1.0.0
 
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @prototype/devtools-core@1.0.0
+  - prototype-devtools-core@1.0.0