+++ packages/devtools-core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @prototype/devtools-core
+# prototype-devtools-core
 
 ## 1.0.0