+++ packages/devtools-overlay/CHANGELOG.md
@@ -9,4 +9,4 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @prototype/devtools-core@1.0.0
+  - prototype-devtools-core@1.0.0