+++ packages/vite-plugin-prototype-devtools/CHANGELOG.md
@@ -10,4 +10,4 @@
 
 - Updated dependencies
   - vite-plugin-prototype@1.0.0
-  - @prototype/devtools-core@1.0.0
+  - prototype-devtools-core@1.0.0